encoders/numeric/UMAPEncoder/requirements.txt
umap-learn==0.5.1